refactor: dto 패키지구조 리팩토링
