+++ src/main/java/com/jscode/board/service/MemberService.java
@@ -1,7 +1,7 @@
 package com.jscode.board.service;
 
 import com.jscode.board.domain.Member;
-import com.jscode.board.dto.request.MemberFormRequest;
+import com.jscode.board.dto.member.MemberFormRequest;
 import com.jscode.board.exception.Member.DuplicateEmailException;
 import com.jscode.board.exception.code.ErrorCode;
 import com.jscode.board.repository.MemberRepository;
